refactor Repositories module
